post e delete

+++ cardapio-digital-api/src/main/java/devweb/cardapiodigitalapi/repository/ComidaRepository.java
@@ -6,8 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import devweb.cardapiodigitalapi.model.Comida;
 
-public interface ComidaRepository extends JpaRepository<Comida, Integer> {
+public interface ComidaRepository extends JpaRepository<Comida, Long> {
 
 	List<Comida> findByNome(String nome);
+	
 
 }
